+++ DmMain/src/Modules/DMLayoutImpl.cpp
@@ -150,7 +150,7 @@ namespace DM
 		{
 			CStringWList strPosList;
 			SplitStringT(strPos,L',',strPosList);
-			m_nCount = strPosList.GetCount();
+			m_nCount = (int)strPosList.GetCount();
 			if (2!=m_nCount&&4!=m_nCount)
 			{
 				break;
