+++ DmMain/src/Widgets/DUIGif.cpp
@@ -366,7 +366,7 @@ namespace DM
 			CStringW strType;
 			CStringW strResName;
 			CStringWList strList;
-			int nCount = SplitStringT(strValue,L':',strList);
+			int nCount = (int)SplitStringT(strValue,L':',strList);
 			if (1== nCount)// ø…ƒ‹ «GIF:IDR_GIF
 			{
 				strType    = L"GIF";
